+++ src/api/java/blusunrize/immersiveengineering/api/tool/LogicCircuitHandler.java
@@ -14,6 +14,7 @@ import com.google.common.base.Preconditions;
 import io.netty.buffer.ByteBuf;
 import malte0811.dualcodecs.DualCodec;
 import malte0811.dualcodecs.DualCompositeCodecs;
+import net.minecraft.ChatFormatting;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.ListTag;
 import net.minecraft.nbt.StringTag;
@@ -84,7 +85,7 @@ public class LogicCircuitHandler
 			return valueOf(name);
 		}
 
-		public Component getDescription()
+		public MutableComponent getDescription()
 		{
 			return Component.translatable(Lib.DESC_INFO+"operator."+this.name().toLowerCase(Locale.ENGLISH));
 		}
@@ -103,7 +104,10 @@ public class LogicCircuitHandler
 		public MutableComponent getDescription()
 		{
 			if(this.ordinal() < 16)
-				return Component.translatable("color.minecraft."+DyeColor.byId(this.ordinal()).getName());
+			{
+				DyeColor dye = DyeColor.byId(this.ordinal());
+				return Component.translatable("color.minecraft."+dye.getName()).withColor(dye.getTextureDiffuseColor());
+			}
 			else
 				return Component.translatable(Lib.DESC_INFO+"register", this.ordinal()-16);
 		}
@@ -133,12 +137,15 @@ public class LogicCircuitHandler
 			this.output = output;
 			this.inputs = inputs;
 
-			this.formattedString = this.output.getDescription();
-			this.formattedString.append(" = ");
-			this.formattedString.append(this.operator.getDescription());
+
+			this.formattedString = Component.empty();
+			this.formattedString.append(this.output.getDescription());
+			this.formattedString.append(
+					Component.literal(" = ").append(this.operator.getDescription()).withStyle(ChatFormatting.GRAY)
+			);
 			for(int i = 0; i < inputs.length; i++)
 			{
-				this.formattedString.append(i!=0?", ": ": ");
+				this.formattedString.append(Component.literal(i!=0?", ": ": ")).withStyle(ChatFormatting.GRAY);
 				this.formattedString.append(inputs[i].getDescription());
 			}
 		}